requirements.txt
aiogram==3.12.0
aiohttp==3.10.5
python-dotenv==1.0.0
openai==1.51.0
certifi==2024.8.30
